Move `calculateTotalBalance` method to utils to reuse it
